Fixed aggpair1sync for real now - with some bykey additions, and var_deps and var_defs. It's kind of a miracle it works, need to sort out and make it coherent.
Add a function for adding new variables including updating the dependencies and definitions (globally). For each expression replaced, we need to just keep a list of the names of the variables created for it, then we should use the global structure to map to actual definitions. Maybe just make it the dependencies locally managed and the definitions global.